+++ src/main/java/com/bbdsoftware/snakey/App.java
@@ -1,0 +1,12 @@
+package com.bbdsoftware.snakey;
+
+import com.bbdsoftware.snakey.presentation.cmdline.Menu;
+
+
+public class App 
+{
+    public static void main(String[] args) {
+        Menu menu = new Menu();
+        menu.newGame();
+    }
+}
