+++ src/main/java/com/bbdsoftware/snakey/domain/Food.java
@@ -1,51 +1,39 @@
-package com.bbd.levelup2023.java.domain;
+package com.bbdsoftware.snakey.domain;
 
-import com.bbd.levelup2023.java.enums.Colour;
+import com.bbdsoftware.snakey.enums.Colour;
 
-public class Apple extends Food{
-    private final String foodName;
-    private final float score;
-    private final Colour foodColour;
-    private final int length;
-    private final Cell foodCell;
+public abstract class Food{
+    private String foodName;
+    private float score;
+    private Colour foodColour;
+    private int length;
+    private Cell foodCell;
 
-    public Apple(Cell food_cell){
-        this.foodName = "Apple";
-        this.score = 0.5F;
-        this.foodColour = Colour.RED;
-        this.length = 2;
+    public Food(){ }
 
-        this.foodCell = food_cell;
-    }
-
-    @Override
     public String getFoodName() {
         return foodName;
     }
 
-    @Override
     public float getScore() {
         return score;
     }
 
-    @Override
     public Colour getFoodColour() {
         return foodColour;
     }
 
-    @Override
     public int getLength() {
         return length;
     }
 
-    @Override
     public Cell getFoodCell() {
         return foodCell;
     }
 
     @Override
     public String toString() {
-        return "Apple{" +
+        return "Food{" +
                 "foodName='" + foodName + '\'' +
                 ", score=" + score +
                 ", foodColour=" + foodColour +
