+++ src/main/java/com/bbdsoftware/snakey/enums/CellType.java
@@ -1,4 +1,4 @@
-package com.bbd.levelup2023.java.enums;
+package com.bbdsoftware.snakey.enums;
 
 public enum CellType {
     SNAKE,
